requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2020.6.20
decorator==5.1.0
Django==3.2
django-filter==21.1
djangorestframework==3.13.0
ipython==7.30.1
jedi==0.18.1
Markdown==3.3.6
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
ptyprocess==0.7.0
Pygments==2.10.0
pytz==2021.3
sqlparse==0.4.2
traitlets==5.1.1
wcwidth==0.2.5
